scraping animations final
