+++ packages/model-noita/src/scraping/noita-gif.ts
@@ -3,7 +3,7 @@ import { SpriteAnimation } from '../common/noita-sprite-animation.ts';
 export interface NoitaGif {
   animationId: string;
   sprite: SpriteAnimation;
-  buffer: Uint8Array;
+  buffer: string;
   repeat: boolean;
   firstFrame: string;
 }
