+++ packages/scrapers/src/scrapers/datawak/scrape-animations/types.ts
@@ -1,0 +1,6 @@
+import { SpriteAnimation } from '@noita-explorer/model-noita';
+
+export interface AnimationFramesResult {
+  animation: SpriteAnimation;
+  frameImages: string[];
+}
