+++ packages/tools/src/common/string.ts
@@ -1,3 +1,5 @@
+import { Buffer } from 'buffer';
+
 const trim = ({
   text,
   fromEnd,
@@ -22,6 +24,11 @@ const trim = ({
   return text;
 };
 
+function uint8ArrayToBase64(uint8Array: Uint8Array): string {
+  return Buffer.from(uint8Array).toString('base64');
+}
+
 export const stringHelpers = {
   trim: trim,
+  uint8ArrayToBase64,
 };
